Backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
torch>=2.1.0 --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.16.0 --index-url https://download.pytorch.org/whl/cpu
ultralytics>=8.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
websockets>=12.0
python-socketio>=5.10.0
aiofiles>=23.0.0
